refactor(frontend): use factories index import
