+++ frontend/__tests__/.server/domain/services/client-friendly-status.service.test.ts
@@ -6,7 +6,7 @@ import type { ClientFriendlyStatusDto } from '~/.server/domain/dtos';
 import type { ClientFriendlyStatusDtoMapper } from '~/.server/domain/mappers';
 import type { ClientFriendlyStatusRepository } from '~/.server/domain/repositories';
 import { ClientFriendlyStatusServiceImpl } from '~/.server/domain/services';
-import type { LogFactory, Logger } from '~/.server/factories/log.factory';
+import type { LogFactory, Logger } from '~/.server/factories';
 
 vi.mock('moize');
