+++ frontend/__tests__/.server/domain/services/country.service.test.ts
@@ -6,7 +6,7 @@ import type { CountryDto } from '~/.server/domain/dtos';
 import type { CountryDtoMapper } from '~/.server/domain/mappers';
 import type { CountryRepository } from '~/.server/domain/repositories';
 import { CountryServiceImpl } from '~/.server/domain/services';
-import type { LogFactory, Logger } from '~/.server/factories/log.factory';
+import type { LogFactory, Logger } from '~/.server/factories';
 
 vi.mock('moize');
