+++ frontend/__tests__/.server/domain/services/marital-status.service.test.ts
@@ -6,7 +6,7 @@ import type { MaritalStatusDto } from '~/.server/domain/dtos';
 import type { MaritalStatusDtoMapper } from '~/.server/domain/mappers';
 import type { MaritalStatusRepository } from '~/.server/domain/repositories';
 import { MaritalStatusServiceImpl } from '~/.server/domain/services';
-import type { LogFactory, Logger } from '~/.server/factories/log.factory';
+import type { LogFactory, Logger } from '~/.server/factories';
 
 vi.mock('moize');
