+++ frontend/app/.server/container-modules/configs.container-module.ts
@@ -2,7 +2,7 @@ import { ContainerModule } from 'inversify';
 
 import type { ClientConfig, ServerConfig } from '~/.server/configs';
 import { SERVICE_IDENTIFIER } from '~/.server/constants';
-import type { ConfigFactory } from '~/.server/factories/config.factory';
+import type { ConfigFactory } from '~/.server/factories';
 
 /**
  * Container module for configurations.
