+++ frontend/app/.server/container-modules/factories.container-module.ts
@@ -1,10 +1,8 @@
 import { ContainerModule } from 'inversify';
 
 import { SERVICE_IDENTIFIER } from '~/.server/constants';
-import { ConfigFactoryImpl } from '~/.server/factories/config.factory';
-import type { ConfigFactory } from '~/.server/factories/config.factory';
-import type { LogFactory } from '~/.server/factories/log.factory';
-import { LogFactoryImpl } from '~/.server/factories/log.factory';
+import { ConfigFactoryImpl , LogFactoryImpl } from '~/.server/factories';
+import type { ConfigFactory , LogFactory } from '~/.server/factories';
 
 /**
  * Container module for factories.
