+++ frontend/app/.server/container.init.ts
@@ -4,8 +4,8 @@ import { makeLoggerMiddleware, textSerializer } from 'inversify-logger-middlewar
 
 import type { ServerConfig } from '~/.server/configs';
 import { SERVICE_IDENTIFIER } from '~/.server/constants';
-import { configsContainerModule , factoriesContainerModule , mappersContainerModule , repositoriesContainerModule , servicesContainerModule } from '~/.server/container-modules';
-import type { Logger } from '~/.server/factories/log.factory';
+import { configsContainerModule, factoriesContainerModule, mappersContainerModule, repositoriesContainerModule, servicesContainerModule } from '~/.server/container-modules';
+import type { Logger } from '~/.server/factories';
 import { getLogger } from '~/utils/logging.server';
 
 export function initContainer() {
