+++ frontend/app/.server/domain/repositories/client-application.repository.ts
@@ -3,7 +3,7 @@ import { inject, injectable } from 'inversify';
 import { ServerConfig } from '~/.server/configs';
 import { SERVICE_IDENTIFIER } from '~/.server/constants';
 import { ClientApplicationEntity } from '~/.server/domain/entities';
-import type { LogFactory, Logger } from '~/.server/factories/log.factory';
+import type { LogFactory, Logger } from '~/.server/factories';
 import { getFetchFn, instrumentedFetch } from '~/utils/fetch-utils.server';
 
 export interface FindByFirstNameLastNameDobClientNumberSearchCriteria {
