+++ frontend/app/.server/domain/repositories/marital-status.repository.ts
@@ -2,7 +2,7 @@ import { inject, injectable } from 'inversify';
 
 import { SERVICE_IDENTIFIER } from '~/.server/constants';
 import type { MaritalStatusEntity } from '~/.server/domain/entities';
-import type { LogFactory, Logger } from '~/.server/factories/log.factory';
+import type { LogFactory, Logger } from '~/.server/factories';
 import maritalStatusJsonDataSource from '~/.server/resources/power-platform/marital-status.json';
 
 export interface MaritalStatusRepository {
