+++ frontend/app/.server/domain/repositories/preferred-language.repository.ts
@@ -2,7 +2,7 @@ import { inject, injectable } from 'inversify';
 
 import { SERVICE_IDENTIFIER } from '~/.server/constants';
 import type { PreferredLanguageEntity } from '~/.server/domain/entities';
-import type { LogFactory, Logger } from '~/.server/factories/log.factory';
+import type { LogFactory, Logger } from '~/.server/factories';
 import preferredLanguageJsonDataSource from '~/.server/resources/power-platform/preferred-language.json';
 
 export interface PreferredLanguageRepository {
