+++ frontend/app/.server/domain/services/client-application.service.ts
@@ -4,7 +4,7 @@ import { SERVICE_IDENTIFIER } from '~/.server/constants';
 import type { ClientApplicationDto } from '~/.server/domain/dtos';
 import type { ClientApplicationDtoMapper } from '~/.server/domain/mappers';
 import type { ClientApplicationRepository } from '~/.server/domain/repositories';
-import type { LogFactory, Logger } from '~/.server/factories/log.factory';
+import type { LogFactory, Logger } from '~/.server/factories';
 
 export interface FindByFirstNameLastNameDobClientNumberSearchCriteria {
   /** The first name of the client. */
